Move the condition block in the addon world
