+++ addons/tutorial_02_connect_on_startup/manifest.json
@@ -2,6 +2,9 @@
   "api_version": "0.1",
   "id": "tutorial_02_connect_on_startup",
   "name": "Tutorial: Connect on startup",
+  "conditions": {
+    "enabledFeatures": ["startOnBoot"]
+  },
   "type": "tutorial",
   "tutorial": {
     "id": "02_connect_on_startup",
@@ -9,9 +12,6 @@
     "title": "Connect VPN on startup",
     "subtitle": "Follow this walkthrough to learn how to activate your VPN when you start your device.",
     "completion_message": "You’ve successfully learned how to set “Connect VPN on startup”. Would you like to learn more tips and tricks?",
-    "conditions": {
-      "enabledFeatures": ["startOnBoot"]
-    },
     "steps": [
       {
         "id": "s1",
