+++ src/addons/addon.cpp
@@ -9,18 +9,115 @@
 #include "addontutorial.h"
 #include "leakdetector.h"
 #include "logger.h"
-#include "models/guidemodel.h"
+#include "models/feature.h"
+#include "mozillavpn.h"
 #include "settingsholder.h"
 
 #include <QCoreApplication>
 #include <QDir>
 #include <QFileInfo>
+#include <QJsonArray>
 #include <QJsonDocument>
 #include <QJsonObject>
 
 namespace {
 Logger logger(LOG_MAIN, "Addon");
+
+bool evaluateConditionsEnabledFeatures(const QJsonArray& enabledFeatures) {
+  for (QJsonValue enabledFeature : enabledFeatures) {
+    QString featureName = enabledFeature.toString();
+
+    // If the feature doesn't exist, we crash.
+    const Feature* feature = Feature::get(featureName);
+    if (!feature) {
+      logger.info() << "Feature not found" << featureName;
+      return false;
+    }
+
+    if (!feature->isSupported()) {
+      logger.info() << "Feature not supported" << featureName;
+      return false;
+    }
+  }
+
+  return true;
 }
+
+bool evaluateConditionsPlatforms(const QJsonArray& platformArray) {
+  QStringList platforms;
+  for (QJsonValue platform : platformArray) {
+    platforms.append(platform.toString());
+  }
+
+  if (!platforms.isEmpty() &&
+      !platforms.contains(MozillaVPN::instance()->platform())) {
+    logger.info() << "Not supported platform";
+    return false;
+  }
+
+  return true;
+}
+
+bool evaluateConditionsSettingsOp(const QString& op, bool result) {
+  if (op == "eq") return result;
+
+  if (op == "neq") return !result;
+
+  logger.warning() << "Invalid settings operator" << op;
+  return false;
+}
+
+bool evaluateConditionsSettings(const QJsonArray& settings) {
+  for (QJsonValue setting : settings) {
+    QJsonObject obj = setting.toObject();
+
+    QString op = obj["op"].toString();
+    QString key = obj["setting"].toString();
+    QVariant valueA = SettingsHolder::instance()->rawSetting(key);
+    if (!valueA.isValid()) {
+      logger.info() << "Unable to retrieve setting key" << key;
+      return false;
+    }
+
+    QJsonValue valueB = obj["value"];
+    switch (valueB.type()) {
+      case QJsonValue::Bool:
+        if (!evaluateConditionsSettingsOp(op,
+                                          valueA.toBool() == valueB.toBool())) {
+          logger.info() << "Setting value doesn't match for key" << key;
+          return false;
+        }
+
+        break;
+
+      case QJsonValue::Double:
+        if (!evaluateConditionsSettingsOp(
+                op, valueA.toDouble() == valueB.toDouble())) {
+          logger.info() << "Setting value doesn't match for key" << key;
+          return false;
+        }
+
+        break;
+        break;
+
+      case QJsonValue::String:
+        if (!evaluateConditionsSettingsOp(
+                op, valueA.toString() == valueB.toString())) {
+          logger.info() << "Setting value doesn't match for key" << key;
+          return false;
+        }
+
+        break;
+
+      default:
+        logger.warning() << "Unsupported setting value type for key" << key;
+        return false;
+    }
+  }
+
+  return true;
+}
+}  // namespace
 
 // static
 Addon* Addon::create(QObject* parent, const QString& manifestFileName) {
@@ -49,6 +146,13 @@ Addon* Addon::create(QObject* parent, const QString& manifestFileName) {
   if (version != "0.1") {
     logger.warning() << "Unsupported API version" << version
                      << manifestFileName;
+    return nullptr;
+  }
+
+  QJsonObject conditions = obj["conditions"].toObject();
+  if (!evaluateConditions(conditions)) {
+    logger.info() << "Exclude the addon because conditions do not match"
+                  << manifestFileName;
     return nullptr;
   }
 
@@ -123,4 +227,22 @@ void Addon::retranslate() {
           QFileInfo(m_manifestFileName).dir().filePath("i18n"))) {
     logger.error() << "Loading the locale failed. - code:" << code;
   }
+}
+
+// static
+bool Addon::evaluateConditions(const QJsonObject& conditions) {
+  if (!evaluateConditionsEnabledFeatures(
+          conditions["enabledFeatures"].toArray())) {
+    return false;
+  }
+
+  if (!evaluateConditionsPlatforms(conditions["platforms"].toArray())) {
+    return false;
+  }
+
+  if (!evaluateConditionsSettings(conditions["settings"].toArray())) {
+    return false;
+  }
+
+  return true;
 }
