+++ src/models/guide.cpp
@@ -6,9 +6,6 @@
 #include "guideblock.h"
 #include "leakdetector.h"
 #include "logger.h"
-#include "models/feature.h"
-#include "mozillavpn.h"
-#include "settingsholder.h"
 
 #include <QFile>
 #include <QJsonArray>
@@ -18,101 +15,6 @@
 
 namespace {
 Logger logger(LOG_MAIN, "Guide");
-
-bool evaluateConditionsEnabledFeatures(const QJsonArray& enabledFeatures) {
-  for (QJsonValue enabledFeature : enabledFeatures) {
-    QString featureName = enabledFeature.toString();
-
-    // If the feature doesn't exist, we crash.
-    const Feature* feature = Feature::get(featureName);
-    if (!feature) {
-      logger.info() << "Feature not found" << featureName;
-      return false;
-    }
-
-    if (!feature->isSupported()) {
-      logger.info() << "Feature not supported" << featureName;
-      return false;
-    }
-  }
-
-  return true;
-}
-
-bool evaluateConditionsPlatforms(const QJsonArray& platformArray) {
-  QStringList platforms;
-  for (QJsonValue platform : platformArray) {
-    platforms.append(platform.toString());
-  }
-
-  if (!platforms.isEmpty() &&
-      !platforms.contains(MozillaVPN::instance()->platform())) {
-    logger.info() << "Not supported platform";
-    return false;
-  }
-
-  return true;
-}
-
-bool evaluateConditionsSettingsOp(const QString& op, bool result) {
-  if (op == "eq") return result;
-
-  if (op == "neq") return !result;
-
-  logger.warning() << "Invalid settings operator" << op;
-  return false;
-}
-
-bool evaluateConditionsSettings(const QJsonArray& settings) {
-  for (QJsonValue setting : settings) {
-    QJsonObject obj = setting.toObject();
-
-    QString op = obj["op"].toString();
-    QString key = obj["setting"].toString();
-    QVariant valueA = SettingsHolder::instance()->rawSetting(key);
-    if (!valueA.isValid()) {
-      logger.info() << "Unable to retrieve setting key" << key;
-      return false;
-    }
-
-    QJsonValue valueB = obj["value"];
-    switch (valueB.type()) {
-      case QJsonValue::Bool:
-        if (!evaluateConditionsSettingsOp(op,
-                                          valueA.toBool() == valueB.toBool())) {
-          logger.info() << "Setting value doesn't match for key" << key;
-          return false;
-        }
-
-        break;
-
-      case QJsonValue::Double:
-        if (!evaluateConditionsSettingsOp(
-                op, valueA.toDouble() == valueB.toDouble())) {
-          logger.info() << "Setting value doesn't match for key" << key;
-          return false;
-        }
-
-        break;
-        break;
-
-      case QJsonValue::String:
-        if (!evaluateConditionsSettingsOp(
-                op, valueA.toString() == valueB.toString())) {
-          logger.info() << "Setting value doesn't match for key" << key;
-          return false;
-        }
-
-        break;
-
-      default:
-        logger.warning() << "Unsupported setting value type for key" << key;
-        return false;
-    }
-  }
-
-  return true;
-}
 
 }  // namespace
 
@@ -140,12 +42,6 @@ Guide* Guide::create(QObject* parent, const QString& fileName) {
 
 // static
 Guide* Guide::create(QObject* parent, const QJsonObject& obj) {
-  QJsonObject conditions = obj["conditions"].toObject();
-  if (!evaluateConditions(conditions)) {
-    logger.info() << "Exclude the guide because conditions do not match";
-    return nullptr;
-  }
-
   QString guideId = obj["id"].toString();
   if (guideId.isEmpty()) {
     logger.warning() << "Empty ID for guide";
@@ -188,22 +84,4 @@ Guide* Guide::create(QObject* parent, const QJsonObject& obj) {
 
   guard.dismiss();
   return guide;
-}
-
-// static
-bool Guide::evaluateConditions(const QJsonObject& conditions) {
-  if (!evaluateConditionsEnabledFeatures(
-          conditions["enabledFeatures"].toArray())) {
-    return false;
-  }
-
-  if (!evaluateConditionsPlatforms(conditions["platforms"].toArray())) {
-    return false;
-  }
-
-  if (!evaluateConditionsSettings(conditions["settings"].toArray())) {
-    return false;
-  }
-
-  return true;
 }
