+++ src/models/tutorialstep.cpp
@@ -3,7 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "tutorialstep.h"
-#include "guide.h"
+#include "addons/addon.h"
 #include "inspector/inspectorutils.h"
 #include "leakdetector.h"
 #include "logger.h"
@@ -106,7 +106,7 @@ void TutorialStep::start() {
 void TutorialStep::startInternal() {
   Q_ASSERT(m_started);
 
-  if (!Guide::evaluateConditions(m_conditions)) {
+  if (!Addon::evaluateConditions(m_conditions)) {
     logger.info()
         << "Exclude the tutorial step because conditions do not match";
     emit completed();
