+++ tests/unit/testaddon.cpp
@@ -1,0 +1,122 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+#include "testaddon.h"
+#include "../../src/addons/addon.h"
+#include "../../src/settingsholder.h"
+#include "helper.h"
+
+void TestAddon::conditions_data() {
+  QTest::addColumn<QJsonObject>("conditions");
+  QTest::addColumn<bool>("result");
+  QTest::addColumn<QString>("settingKey");
+  QTest::addColumn<QVariant>("settingValue");
+
+  QTest::addRow("empty") << QJsonObject() << true << "" << QVariant();
+
+  {
+    QJsonObject obj;
+    obj["platforms"] = QJsonArray{"foo"};
+    QTest::addRow("platforms") << obj << false << "" << QVariant();
+  }
+
+  {
+    QJsonObject obj;
+    obj["enabledFeatures"] = QJsonArray{"appReview"};
+    QTest::addRow("enabledFeatures") << obj << false << "" << QVariant();
+  }
+
+  {
+    QJsonObject obj;
+    QJsonObject settings;
+    obj["settings"] = QJsonArray{settings};
+    QTest::addRow("empty settings") << obj << false << "" << QVariant();
+
+    settings["op"] = "eq";
+    settings["setting"] = "foo";
+    settings["value"] = true;
+    obj["settings"] = QJsonArray{settings};
+    QTest::addRow("invalid settings") << obj << false << "" << QVariant();
+
+    QTest::addRow("string to boolean type settings - boolean")
+        << obj << true << "foo" << QVariant("wow");
+
+    QTest::addRow("op=eq settings - boolean")
+        << obj << true << "foo" << QVariant(true);
+
+    QTest::addRow("op=eq settings - boolean 2")
+        << obj << false << "foo" << QVariant(false);
+
+    settings["op"] = "WOW";
+    obj["settings"] = QJsonArray{settings};
+    QTest::addRow("invalid op settings - boolean")
+        << obj << false << "foo" << QVariant(false);
+
+    settings["op"] = "neq";
+    obj["settings"] = QJsonArray{settings};
+    QTest::addRow("op=neq settings - boolean")
+        << obj << true << "foo" << QVariant(false);
+
+    settings["op"] = "eq";
+    settings["value"] = 42;
+    obj["settings"] = QJsonArray{settings};
+    QTest::addRow("invalid type settings - double")
+        << obj << false << "foo" << QVariant("wow");
+
+    QTest::addRow("op=eq settings - double")
+        << obj << true << "foo" << QVariant(42);
+
+    QTest::addRow("op=eq settings - double 2")
+        << obj << false << "foo" << QVariant(43);
+
+    settings["op"] = "WOW";
+    obj["settings"] = QJsonArray{settings};
+    QTest::addRow("invalid op settings - double")
+        << obj << false << "foo" << QVariant(43);
+
+    settings["op"] = "neq";
+    obj["settings"] = QJsonArray{settings};
+    QTest::addRow("op=neq settings - double")
+        << obj << true << "foo" << QVariant(43);
+
+    settings["op"] = "eq";
+    settings["value"] = "wow";
+    obj["settings"] = QJsonArray{settings};
+    QTest::addRow("invalid type settings - string")
+        << obj << false << "foo" << QVariant(false);
+
+    QTest::addRow("op=eq settings - string")
+        << obj << true << "foo" << QVariant("wow");
+
+    QTest::addRow("op=eq settings - string 2")
+        << obj << false << "foo" << QVariant("wooow");
+
+    settings["op"] = "WOW";
+    obj["settings"] = QJsonArray{settings};
+    QTest::addRow("invalid op settings - string")
+        << obj << false << "foo" << QVariant("woow");
+
+    settings["op"] = "neq";
+    obj["settings"] = QJsonArray{settings};
+    QTest::addRow("op=neq settings - string")
+        << obj << true << "foo" << QVariant("woow");
+  }
+}
+
+void TestAddon::conditions() {
+  SettingsHolder settingsHolder;
+
+  QFETCH(QJsonObject, conditions);
+  QFETCH(bool, result);
+  QFETCH(QString, settingKey);
+  QFETCH(QVariant, settingValue);
+
+  if (!settingKey.isEmpty()) {
+    settingsHolder.setRawSetting(settingKey, settingValue);
+  }
+
+  QCOMPARE(Addon::evaluateConditions(conditions), result);
+}
+
+static TestAddon s_testAddon;
