+++ tests/unit/testaddon.h
@@ -4,13 +4,10 @@
 
 #include "helper.h"
 
-class TestGuide final : public TestHelper {
+class TestAddon final : public TestHelper {
   Q_OBJECT
 
  private slots:
-  void create_data();
-  void create();
-  void createNotExisting();
-
-  void model();
+  void conditions_data();
+  void conditions();
 };
