+++ tests/unit/testguide.cpp
@@ -104,16 +104,6 @@ void TestGuide::create_data() {
   obj["blocks"] = blocks;
   QTest::addRow("with-block-type-olist-with-subblock")
       << "foo" << QJsonDocument(obj).toJson() << true;
-
-  obj["conditions"] = QJsonObject();
-  QTest::addRow("with-block-type-olist-with-subblock and conditions")
-      << "foo" << QJsonDocument(obj).toJson() << true;
-
-  block["type"] = "ulist";
-  blocks.replace(0, block);
-  obj["conditions"] = QJsonObject();
-  QTest::addRow("with-block-type-ulist-with-subblock and conditions")
-      << "foo" << QJsonDocument(obj).toJson() << true;
 }
 
 void TestGuide::create() {
@@ -173,118 +163,6 @@ void TestGuide::model() {
   Guide* guide =
       mg->data(mg->index(0, 0), GuideModel::GuideRole).value<Guide*>();
   QVERIFY(!!guide);
-}
-
-void TestGuide::conditions_data() {
-  QTest::addColumn<QJsonObject>("conditions");
-  QTest::addColumn<bool>("result");
-  QTest::addColumn<QString>("settingKey");
-  QTest::addColumn<QVariant>("settingValue");
-
-  QTest::addRow("empty") << QJsonObject() << true << "" << QVariant();
-
-  {
-    QJsonObject obj;
-    obj["platforms"] = QJsonArray{"foo"};
-    QTest::addRow("platforms") << obj << false << "" << QVariant();
-  }
-
-  {
-    QJsonObject obj;
-    obj["enabledFeatures"] = QJsonArray{"appReview"};
-    QTest::addRow("enabledFeatures") << obj << false << "" << QVariant();
-  }
-
-  {
-    QJsonObject obj;
-    QJsonObject settings;
-    obj["settings"] = QJsonArray{settings};
-    QTest::addRow("empty settings") << obj << false << "" << QVariant();
-
-    settings["op"] = "eq";
-    settings["setting"] = "foo";
-    settings["value"] = true;
-    obj["settings"] = QJsonArray{settings};
-    QTest::addRow("invalid settings") << obj << false << "" << QVariant();
-
-    QTest::addRow("string to boolean type settings - boolean")
-        << obj << true << "foo" << QVariant("wow");
-
-    QTest::addRow("op=eq settings - boolean")
-        << obj << true << "foo" << QVariant(true);
-
-    QTest::addRow("op=eq settings - boolean 2")
-        << obj << false << "foo" << QVariant(false);
-
-    settings["op"] = "WOW";
-    obj["settings"] = QJsonArray{settings};
-    QTest::addRow("invalid op settings - boolean")
-        << obj << false << "foo" << QVariant(false);
-
-    settings["op"] = "neq";
-    obj["settings"] = QJsonArray{settings};
-    QTest::addRow("op=neq settings - boolean")
-        << obj << true << "foo" << QVariant(false);
-
-    settings["op"] = "eq";
-    settings["value"] = 42;
-    obj["settings"] = QJsonArray{settings};
-    QTest::addRow("invalid type settings - double")
-        << obj << false << "foo" << QVariant("wow");
-
-    QTest::addRow("op=eq settings - double")
-        << obj << true << "foo" << QVariant(42);
-
-    QTest::addRow("op=eq settings - double 2")
-        << obj << false << "foo" << QVariant(43);
-
-    settings["op"] = "WOW";
-    obj["settings"] = QJsonArray{settings};
-    QTest::addRow("invalid op settings - double")
-        << obj << false << "foo" << QVariant(43);
-
-    settings["op"] = "neq";
-    obj["settings"] = QJsonArray{settings};
-    QTest::addRow("op=neq settings - double")
-        << obj << true << "foo" << QVariant(43);
-
-    settings["op"] = "eq";
-    settings["value"] = "wow";
-    obj["settings"] = QJsonArray{settings};
-    QTest::addRow("invalid type settings - string")
-        << obj << false << "foo" << QVariant(false);
-
-    QTest::addRow("op=eq settings - string")
-        << obj << true << "foo" << QVariant("wow");
-
-    QTest::addRow("op=eq settings - string 2")
-        << obj << false << "foo" << QVariant("wooow");
-
-    settings["op"] = "WOW";
-    obj["settings"] = QJsonArray{settings};
-    QTest::addRow("invalid op settings - string")
-        << obj << false << "foo" << QVariant("woow");
-
-    settings["op"] = "neq";
-    obj["settings"] = QJsonArray{settings};
-    QTest::addRow("op=neq settings - string")
-        << obj << true << "foo" << QVariant("woow");
-  }
-}
-
-void TestGuide::conditions() {
-  SettingsHolder settingsHolder;
-
-  QFETCH(QJsonObject, conditions);
-  QFETCH(bool, result);
-  QFETCH(QString, settingKey);
-  QFETCH(QVariant, settingValue);
-
-  if (!settingKey.isEmpty()) {
-    settingsHolder.setRawSetting(settingKey, settingValue);
-  }
-
-  QCOMPARE(Guide::evaluateConditions(conditions), result);
 }
 
 static TestGuide s_testGuide;
